requirements.txt
opencv_python==4.9.0.80
Pillow==10.2.0
selenium==4.18.1
